requirements.txt
pyarrow==11.0.0
pylint==2.16.2
pytest-cov==4.0.0
pytest-dotenv==0.5.2
pytest==7.2.1
python-dotenv==1.0.0
pymongo==4.3.3
